Swap Upgrade Screen
Pop up screen contains 5 buttons with the current upgrades that can be swapped out.
Pops up when player has 5 upgrades and player selects another upgrade. Currently no swap functionality, will connect switching the HUD image once there is backend functionality, as it is based on the list in the backend.