Starting a lot of things! Made new screens for enemies and upgrade systems, added a goblin image, started a goblin information page
Going to be going with a tab system where on the side the categories open up images of the compendium pages but stay visible. Clicking one will close all and open another.
Currently only Goblin is set up, and only the instance to open it is working, have not worked on closing it yet.